fix: Change Ids types, feat: Add database AutoFeel

+++ Back/MishFit/MishFit/Services/MealsService.cs
@@ -1,5 +1,6 @@
 ﻿using MishFit.Contracts;
 using MishFit.Entities;
+using MishFit.Exceptions;
 using MishFit.Repositories;
 
 namespace MishFit.Services;
@@ -18,8 +19,10 @@ public class MealsService : IMealsService
         return await _repository.GetAllMealsAsync();
     }
 
-    public async Task<Meal> GetMealByIdAsync(Guid id)
+    public async Task<Meal> GetMealByIdAsync(long id)
     {
+        if (id <= 0)
+            throw new InvalidIncomingParameterException($"Meal id cannot be null.");
         return await _repository.GetMealByIdAsync(id);
     }
 
@@ -33,8 +36,11 @@ public class MealsService : IMealsService
         return await _repository.UpdateMealAsync(contract);
     }
 
-    public async Task<Meal> DeleteMealByIdAsync(Guid id)
+    public async Task<Meal> DeleteMealByIdAsync(long id)
     {
+        if (id <= 0)
+            throw new InvalidIncomingParameterException($"Meal id cannot be null.");
+
         return await _repository.DeleteMealByIdAsync(id);
     }
 }
